factor out place search functionality, add new module for place landing
