+++ static/js/place/search.ts
@@ -1,0 +1,61 @@
+/**
+ * Copyright 2020 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import axios from "axios";
+
+let ac: google.maps.places.Autocomplete;
+
+/**
+ * Setup search input autocomplete
+ */
+function initSearchAutocomplete() {
+  // Create the autocomplete object, restricting the search predictions to
+  // geographical location types.
+  const options = {
+    types: ["(regions)"],
+    fields: ["place_id", "name", "types"],
+  };
+  const acElem = document.getElementById(
+    "place-autocomplete"
+  ) as HTMLInputElement;
+  ac = new google.maps.places.Autocomplete(acElem, options);
+  ac.addListener("place_changed", getPlaceAndRender);
+}
+
+/*
+ * Get place from autocomplete object and update url
+ */
+function getPlaceAndRender() {
+  // Get the place details from the autocomplete object.
+  const place = ac.getPlace();
+  axios
+    .get(`api/placeid2dcid/${place.place_id}`)
+    .then((resp) => {
+      const urlParams = new URLSearchParams(window.location.search);
+      urlParams.set("dcid", resp.data);
+      window.location.search = urlParams.toString();
+    })
+    .catch(() => {
+      alert("Sorry, but we don't have any data about " + place.name);
+      const acElem = document.getElementById(
+        "place-autocomplete"
+      ) as HTMLInputElement;
+      acElem.value = "";
+      acElem.setAttribute("placeholder", "Search for another place");
+    });
+}
+
+export { initSearchAutocomplete };
